Refactored code and package. Added documentation to the methods
setup.py created and initialised with the values.
requirements.txt created and added the initial requirement of iexfinance.
tests folder added with 2 files for unit and integration testing.
DataPuller and DataPusher classes created as a wrapper of the data_pull and data_push methods.
DataOrchestrator added as orchestrator of the data pull and push pipeline.